Initial commit
This code was copied from the `test-util/` directories in the hypothesis/via and
hypothesis/lms repos. The code in `wait.ts` was modified to have two separate
timeouts for warning about slow tests vs actually failing the test. This
encourages developers to fix slow waits, without causing tests to be flakey if
the system is running slow for any reason.